chore: keep all types in one file to avoid misleading
