--- src/types/plugins.d.ts
@@ -1,5 +1,0 @@
-import { BaseSchemaReturns } from "./schema";
-
-export type PluginFunction = (
-  baseReturns: BaseSchemaReturns
-) => BaseSchemaReturns;
